+++ src/components/itineraries/IndividualTripDetails.js
@@ -3,6 +3,8 @@ import { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 import { ActivitySchedule } from "./ActivitySchedule";
 import "./IndividualTripDetails.css";
+import { TextInput, Button, Timeline } from "@mantine/core";
+import { MultiSelect } from "@mantine/core";
 import { getAllActivities } from "../api/APImanager";
 
 //
@@ -163,18 +165,22 @@ export const IndividualTripDetails = ({}) => {
 
   const renderDeleteItineraryButton = () => {
     return (
-      <button
+      <Button
+        color="red"
         onClick={() => {
-          fetch(`http://localhost:8099/itineraries/${foundItinerary.id}`, {
-            method: "DELETE",
-          }).then(() => {
-            navigate("/trips");
-          });
+          if (window.confirm("Press OK to confirm your delete.")) {
+            fetch(`http://localhost:8099/itineraries/${foundItinerary.id}`, {
+              method: "DELETE",
+            }).then(() => {
+              navigate("/trips");
+            });
+          } else {
+            return "";
+          }
         }}
-        className="deletebutton"
       >
         Delete Trip
-      </button>
+      </Button>
     );
   };
 
@@ -202,7 +208,11 @@ export const IndividualTripDetails = ({}) => {
     return (
       <button
         onClick={(event) => {
-          postActivity(event);
+          if (activity.name === "") {
+            window.alert("add text to add an activity");
+          } else {
+            postActivity(event);
+          }
         }}
       >
         Add activity!
@@ -212,210 +222,224 @@ export const IndividualTripDetails = ({}) => {
 
   return (
     <>
-      <section className="full-trip-view">
-        <section className={destination?.location.city + "tripheader"}>
-          <div className="headeroverlay">
-            <h2>
-              Traveling to {destination?.location?.city},{" "}
-              {destination?.location?.state}
-            </h2>
-            <div className="accessCode">
-              Access code: {userItinerary?.itinerary?.accessCode}
+      <main>
+        <section className="full-trip-view">
+          <section className={destination?.location.city + "tripheader"}>
+            <div className="headeroverlay">
+              <h2>
+                Traveling to {destination?.location?.city},{" "}
+                {destination?.location?.state}
+              </h2>
+              <div className="accessCode">
+                Access code: {userItinerary?.itinerary?.accessCode}
+              </div>
+              <Link
+                className="headereditlink"
+                to={`/trips/${userItinerary?.itineraryId}/editItinerary`}
+              >
+                Edit your Itinerary
+              </Link>
+              {renderDeleteItineraryButton()}
+              <h3 class="departuredate">
+                Leaving on: {userItinerary?.itinerary?.departureDate}
+              </h3>
+              <h3 class="returndate">
+                Returning on: {userItinerary?.itinerary?.returnDate}
+              </h3>
             </div>
-            <Link
-              className="headereditlink"
-              to={`/trips/${userItinerary?.itineraryId}/editItinerary`}
-            >
-              Edit your Itinerary
-            </Link>
-            {renderDeleteItineraryButton()}
-            <h3 class="departuredate">
-              Leaving on: {userItinerary?.itinerary?.departureDate}
-            </h3>
-            <h3 class="returndate">
-              Returning on: {userItinerary?.itinerary?.returnDate}
-            </h3>
+          </section>
+          <div className="transportationinfo">
+            <section className="flightinfo">
+              <h4>Flight Info:</h4>
+              <div>
+                Flight to {destination?.location?.city}:{" "}
+                {userItinerary?.itinerary?.flightInfo?.departFlightNum} leaving
+                at:{" "}
+                {userItinerary?.itinerary?.flightInfo?.flightToDestinationTime}
+              </div>
+              <div>
+                Return Flight:{" "}
+                {userItinerary?.itinerary?.flightInfo?.returnFlightNum} leaving
+                at: {userItinerary?.itinerary?.flightInfo?.returnFlightTime}{" "}
+              </div>
+            </section>
+            <section className="rentalinfo">
+              <h4>Rental Car Info:</h4>
+              <div>
+                Renting from:
+                {userItinerary?.itinerary?.rentalCarInfo?.rentalCompany}
+              </div>
+              <div>
+                Reservation #:{" "}
+                {userItinerary?.itinerary?.rentalCarInfo?.reservationNum}
+              </div>
+              <div>
+                Reservation start:{" "}
+                {userItinerary?.itinerary?.rentalCarInfo?.reservationTime}
+              </div>
+              <div>
+                Car Drop off Time:{" "}
+                {userItinerary?.itinerary?.rentalCarInfo?.carDropOffTime}
+              </div>
+            </section>
           </div>
         </section>
-        <div className="transportationinfo">
-          <section className="flightinfo">
-            <h4>Flight Info:</h4>
-            <div>
-              Flight to {destination?.location?.city}:{" "}
-              {userItinerary?.itinerary?.flightInfo?.departFlightNum} leaving
-              at:{" "}
-              {userItinerary?.itinerary?.flightInfo?.flightToDestinationTime}
-            </div>
-            <div>
-              Return Flight:{" "}
-              {userItinerary?.itinerary?.flightInfo?.returnFlightNum} leaving
-              at: {userItinerary?.itinerary?.flightInfo?.returnFlightTime}{" "}
-            </div>
+        <div className="scheduleandactivities">
+          <section className="schedulecontainer">
+            <h3>My Schedule</h3>
+            <section className="schedulelist">
+              <Timeline color="violet" lineWidth={1}>
+                {itineraryActivities.map((itineraryActivity) => (
+                  <ActivitySchedule
+                    key={`itineraryActivity--${itineraryActivity?.id}`}
+                    itineraryActivityObject={itineraryActivity}
+                    activity={itineraryActivity?.activity?.name}
+                    activityDescription={itineraryActivity?.description}
+                    activityAddress={itineraryActivity?.address}
+                    activityDateTime={itineraryActivity?.activityDateTime}
+                    setItineraryActivities={setItineraryActivities}
+                    id={itineraryActivity?.id}
+                    itineraryId={itineraryActivity?.itineraryId}
+                  />
+                ))}
+              </Timeline>
+            </section>
           </section>
-          <section className="rentalinfo">
-            <h4>Rental Car Info:</h4>
-            <div>
-              Renting from:
-              {userItinerary?.itinerary?.rentalCarInfo?.rentalCompany}
-            </div>
-            <div>
-              Reservation #:{" "}
-              {userItinerary?.itinerary?.rentalCarInfo?.reservationNum}
-            </div>
-            <div>
-              Reservation start:{" "}
-              {userItinerary?.itinerary?.rentalCarInfo?.reservationTime}
-            </div>
-            <div>
-              Car Drop off Time:{" "}
-              {userItinerary?.itinerary?.rentalCarInfo?.carDropOffTime}
-            </div>
-          </section>
-        </div>
-      </section>
-      <section>
-        <h3>Activities</h3>
-        <button
-          onClick={() => {
-            setFormVisibility(true);
-          }}
-        >
-          Display Activity Form
-        </button>
-        <button
-          onClick={() => {
-            setFormVisibility(false);
-          }}
-        >
-          Hide Activity Form
-        </button>
-      </section>
-      {activityFormVisibility ? (
-        <>
-          <section>
-            <form>
-              <fieldset>
-                <div className="form-group">
-                  <label htmlFor="name">Choose Activity:</label>
-                  <select
-                    className="form-control"
-                    value={itineraryActivity.activityId}
-                    required
-                    autoFocus
-                    onChange={(evt) => {
-                      const copy = { ...itineraryActivity };
-                      copy.activityId = evt.target.value;
-                      updateItineraryActivity(copy);
-                    }}
-                  >
-                    <option value="0">Choose your Activity</option>
-                    {activities.map((activity) => {
-                      return (
-                        <option value={activity.id} key={activity.id}>
-                          {activity.name}
-                        </option>
-                      );
-                    })}
-                  </select>
-                </div>
-              </fieldset>
-              <fieldset>
-                <div className="departure">
-                  <label htmlFor="description">Add Activity:</label>
-                  <input
-                    required
-                    autoFocus
-                    className="form-control"
-                    type="text"
-                    value={activity.name}
-                    onChange={(evt) => {
-                      const copy = { ...activity };
-                      copy.name = evt.target.value;
-                      updateActivity(copy);
-                    }}
-                  />
-                </div>
-                {addActivityButton()}
-              </fieldset>
-              <fieldset>
-                <div className="departure">
-                  <label htmlFor="description">Description:</label>
-                  <input
-                    autoFocus
-                    className="form-control"
-                    type="text"
-                    value={itineraryActivity.description}
-                    onChange={(evt) => {
-                      const copy = { ...itineraryActivity };
-                      copy.description = evt.target.value;
-                      updateItineraryActivity(copy);
-                    }}
-                  />
-                </div>
-              </fieldset>
-              <fieldset>
-                <div className="departure">
-                  <label htmlFor="description">Activity Time and Date:</label>
-                  <input
-                    autoFocus
-                    className="form-control"
-                    type="datetime-local"
-                    value={itineraryActivity.activityDateTime}
-                    onChange={(evt) => {
-                      const copy = { ...itineraryActivity };
-                      copy.activityDateTime = evt.target.value;
-                      updateItineraryActivity(copy);
-                    }}
-                  />
-                </div>
-              </fieldset>
-              <fieldset>
-                <div className="departure">
-                  <label htmlFor="description">Activity Address:</label>
-                  <input
-                    autoFocus
-                    className="form-control"
-                    type="text"
-                    value={itineraryActivity.address}
-                    onChange={(evt) => {
-                      const copy = { ...itineraryActivity };
-                      copy.address = evt.target.value;
-                      updateItineraryActivity(copy);
-                    }}
-                  />
-                </div>
-              </fieldset>
-              <fieldset></fieldset>
-              <button
-                onClick={(clickEvent) => postItineraryActivity(clickEvent)}
-                className="btn btn-primary"
+          <div class="activitiesform">
+            <section>
+              <h3>Activities</h3>
+              <Button
+                onClick={() => {
+                  setFormVisibility(true);
+                }}
               >
-                Add to Itinerary!
-              </button>
-            </form>
-          </section>
-        </>
-      ) : (
-        ""
-      )}
-      <section className="schedulecontainer">
-        <h3>My Schedule</h3>
-        <section className="schedulelist">
-          {itineraryActivities.map((itineraryActivity) => (
-            <ActivitySchedule
-              key={`itineraryActivity--${itineraryActivity?.id}`}
-              itineraryActivityObject={itineraryActivity}
-              activity={itineraryActivity?.activity?.name}
-              activityDescription={itineraryActivity?.description}
-              activityAddress={itineraryActivity?.address}
-              activityDateTime={itineraryActivity?.activityDateTime}
-              setItineraryActivities={setItineraryActivities}
-              id={itineraryActivity?.id}
-              itineraryId={itineraryActivity?.itineraryId}
-            />
-          ))}
-        </section>
-      </section>
+                Display Activity Form
+              </Button>
+              <Button
+                onClick={() => {
+                  setFormVisibility(false);
+                }}
+              >
+                Hide Activity Form
+              </Button>
+            </section>
+            {activityFormVisibility ? (
+              <>
+                <section>
+                  <form>
+                    <div className="row g-3">
+                      <fieldset>
+                        <div className="col-md-6">
+                          <label htmlFor="name">Choose Activity:</label>
+                          <select
+                            value={itineraryActivity.activityId}
+                            required
+                            autoFocus
+                            onChange={(evt) => {
+                              const copy = { ...itineraryActivity };
+                              copy.activityId = evt.target.value;
+                              updateItineraryActivity(copy);
+                            }}
+                          >
+                            <option value="0">Choose your Activity</option>
+                            {activities.map((activity) => {
+                              return (
+                                <option value={activity.id} key={activity.id}>
+                                  {activity.name}
+                                </option>
+                              );
+                            })}
+                          </select>
+                        </div>
+                      </fieldset>
+                      <fieldset>
+                        <div className="col-md-6">
+                          <label htmlFor="description">Add Activity:</label>
+                          <input
+                            required
+                            autoFocus
+                            className="form-control"
+                            type="text"
+                            value={activity.name}
+                            onChange={(evt) => {
+                              const copy = { ...activity };
+                              copy.name = evt.target.value;
+                              updateActivity(copy);
+                            }}
+                          />
+                        </div>
+                        {addActivityButton()}
+                      </fieldset>
+                    </div>
+                    <fieldset>
+                      <div>
+                        <TextInput
+                          autoFocus
+                          label="Description"
+                          description="Describe the Activity"
+                          type="text"
+                          width={300}
+                          value={itineraryActivity.description}
+                          onChange={(evt) => {
+                            const copy = { ...itineraryActivity };
+                            copy.description = evt.target.value;
+                            updateItineraryActivity(copy);
+                          }}
+                        />
+                      </div>
+                    </fieldset>
+                    <fieldset>
+                      <div className="departure">
+                        <label htmlFor="description">
+                          Activity Time and Date:
+                        </label>
+                        <input
+                          autoFocus
+                          className="form-control"
+                          type="datetime-local"
+                          value={itineraryActivity.activityDateTime}
+                          onChange={(evt) => {
+                            const copy = { ...itineraryActivity };
+                            copy.activityDateTime = evt.target.value;
+                            updateItineraryActivity(copy);
+                          }}
+                        />
+                      </div>
+                    </fieldset>
+                    <fieldset>
+                      <div>
+                        <TextInput
+                          autoFocus
+                          description="Where is this activity located?"
+                          label="Activity Address"
+                          type="text"
+                          width={300}
+                          value={itineraryActivity.address}
+                          onChange={(evt) => {
+                            const copy = { ...itineraryActivity };
+                            copy.address = evt.target.value;
+                            updateItineraryActivity(copy);
+                          }}
+                        />
+                      </div>
+                    </fieldset>
+                    <Button
+                      onClick={(clickEvent) =>
+                        postItineraryActivity(clickEvent)
+                      }
+                      className="btn btn-primary"
+                    >
+                      Add to Itinerary!
+                    </Button>
+                  </form>
+                </section>
+              </>
+            ) : (
+              ""
+            )}
+          </div>
+        </div>
+      </main>
     </>
   );
 };
